+++ src/main/java/com/example/kiosk/manager/service/ManagerService.java
@@ -9,8 +9,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import static com.example.kiosk.global.type.ErrorCode.EXIST_ACCOUNT_EMAIL;
-import static com.example.kiosk.global.type.ErrorCode.NOT_FOUND_ID;
+import static com.example.kiosk.global.type.ErrorCode.*;
 
 @RequiredArgsConstructor
 @Service
@@ -46,6 +45,11 @@ public class ManagerService {
         // 매니저 아이디 번호
         Manager manager = getManagerId(id);
 
+        // 계정 삭제 검사
+        if (Boolean.TRUE.equals(manager.getDeletedYn())) {
+            throw new ManagerException(ALREADY_DELETE_MANAGER);
+        }
+
         manager.deleteManager();
         return manager;
     }
@@ -53,13 +57,13 @@ public class ManagerService {
     // 매니저 아이디 번호
     private Manager getManagerId(Long id) {
         return managerRepository.findById(id)
-                .orElseThrow(() -> new ManagerException(NOT_FOUND_ID));
+                .orElseThrow(() -> new ManagerException(NOT_FOUND_MANAGER_ID));
     }
 
     // 중복 이메일 검사
     private void existManagerEmail(String email) {
         if (managerRepository.findByEmail(email).isPresent()) {
-            throw new ManagerException(EXIST_ACCOUNT_EMAIL);
+            throw new ManagerException(EXIST_MANAGER_EMAIL);
         }
     }
 }
